requirements.txt
adal==1.2.7
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
azure-cognitiveservices-speech==1.31.0
azure-core==1.29.2
azure-functions==1.17.0
azure-identity==1.14.0
azure-storage-blob==12.17.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
cssselect==1.2.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
dnspython==2.4.2
docopt==0.6.2
executing==1.2.0
extra-streamlit-components==0.1.56
fake-useragent==1.2.1
filetype==1.2.0
Flask==2.3.2
Flask-Session==0.5.0
frozendict==2.3.8
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.85.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
html2docx==1.6.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.281
langsmith==0.0.23
lark==1.1.7
loguru==0.7.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==3.1.0
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.23.5
openai==0.27.9
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
parse==1.19.1
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.2
pinecone-text==0.5.2
pipreqs==0.4.13
platformdirs==3.10.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.0
pyee==8.2.2
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-html==0.10.0
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data==0.6.2
streamlit==1.26.0
streamlit-authenticator==0.2.2
streamlit-feedback==0.0.9
streamlit-javascript==0.1.5
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.3.3
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.9.2
uritemplate==4.1.1
urllib3==1.26.16
validators==0.21.2
w3lib==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.7
wget==3.2
win32-setctime==1.1.0
yarg==0.1.9
yarl==1.9.2
yfinance==0.2.27
zipp==3.16.2
